Refactor the DynamicFlow
